New bbcp_C32C class
This new class implements the CRC32c checksum both in hardware (when
available) and in software.

The hardware implementation is based on code from the crc32c python
package (which we author), while the software implementation is based on
the Slicing-by-8 algorithm published by Intel (references are in the
code).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>