Fix filesystem checks for embedded mount points
The bbcp_FS_Unix class uses the path passed down via Applicable() when
checking for available space in the underlying system. However, this
path was being trimmed to its first component (e.g., "/mnt" in
"/mnt/a/b/c/"), incorrectly assuming that all filesystems in a (Unix)
computer are mounted at the root level.

This assumption breaks at least in two cases: firstly, one can have
"/mnt/submount" mounted over "/mnt", which itself can be a mount point.
And secondly, one can have a "/mnt/a" symbolic link pointing to "/a",
which can be a different filesystem.

The solution is simple: don't strip the incoming path from any of its
components, and use it "as is" later during statvfs.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>