Implementing logic for new -j option
With this change users can now make sure bbcp uses the hostname or IP
embedded into the source and/or target file specifications in the
command line arguments. For example, with `bbcp -j
192.168.0.1:/source/file.txt 192.168.0.2:/target/file.txt` the copy of
bbcp running in 192.168.0.1 will connect to 192.168.0.2 regardless of
that host's hostname, and which address that hostname resolves to. This
is useful for setups where two computers have more than one to reach
each other, with one link being preferred over the other.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>